+++ lib/rubocop/cop/lint/uri_escape_unescape.rb
@@ -17,7 +17,8 @@ module RuboCop
       #
       #   # good
       #   CGI.escape('http://example.com')
-      #   URI.encode_www_form('http://example.com')
+      #   URI.encode_www_form([['example', 'param'], ['lang', 'en']])
+      #   URI.encode_www_form(page: 10, locale: 'en')
       #   URI.encode_www_form_component('http://example.com')
       #
       #   # bad